Fix a goal pos tracking error
This fixes a bug where the goal position was never cleared, leading
to action loops. We also figured out what causes the bug where the
robot doesn't claim every cone. Essential, the robot gets stuck in
a loop of placing items on the junction at (2,3). This happens until
we run out of cones. More investigation is required.